delete_materialapp_fix_design
