+++ qiita_app/lib/constants.dart
@@ -1,0 +1,1 @@
+//アプリケーションの色設定やテーマ設定
