+++ qiita_app/lib/pages/my_page.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 class MyPage extends StatefulWidget {
+
   const MyPage({Key? key}) : super(key: key);
 
   @override
